CDP #163 - Extracting Timeline to own component; Updating index.js to export Timeline
